[move]SeaAreaクラスをlogbook.constantsパッケージへ移動
